update: method updateStudent

+++ Colegio/controller/ControllerStudents.php
@@ -5,6 +5,7 @@ $method = $_SERVER['REQUEST_METHOD'];
 $gestor = new GestionEstudiante();
 switch ($method) {
     case 'POST':
+        $action = $_GET['action'];
         $nombre1 = $_POST['nombre1'];
         $nombre2 = $_POST['nombre2'];
         $apellido1 = $_POST['apellido1'];
@@ -12,49 +13,44 @@ switch ($method) {
         $direccion = $_POST['direccion'];
         $telefono = $_POST['telefono'];
         $correo = $_POST['correo'];
-        $estudiante = new Estudiante(0, $nombre1, $nombre2, $apellido1, $apellido2, $direccion, $telefono, $correo, true);
-        try {
-            $gestor->insert($estudiante);
-            echo "<script type='text/javascript'>alert('Registro de estudiante exitoso.');</script>";
-        } catch (Exception $e) {
-            echo "<script type='text/javascript'>console.log(`Error: $e`);</script>";
-            echo "<script type='text/javascript'>alert('No se pudo realizar el registro, lo sentimos.');</script>";
-        }
-        echo "<script type='text/javascript'>window.location.href='../view/home.php';</script>";
-        break;
-    case 'GET':
-        $action = $_GET['action'];
         switch ($action) {
-            case 'editStudent':
-                $id = $_GET['id'];
+            case 'insert':
+                $estudiante = new Estudiante(0, $nombre1, $nombre2, $apellido1, $apellido2, $direccion, $telefono, $correo, true);
                 try {
-                    $estudiante = $gestor->search($id);
-                    if($estudiante){
-                       session_start();
-                       $_SESSION['student'] = $estudiante;
-                        echo "
-                        <script src='../assets/js/ControllerPages.js'></script>
-                          <script type='text/javascript'>
-                            loadPage('../view/edit-student');
-                          </script>";
-                    }else{
-                        echo "<script type='text/javascript'>alert('Estudiante no encontrado.');</script>";
-                        echo "<script type='text/javascript'>window.location.href='../view/home.php';</script>";
-                    }
-                } catch (Exception $e) {
-                    echo "<script type='text/javascript'>alert('No se pudo realizar la búsqueda.');</script>";
-                }
-                break;
-            case 'deleteStudent':
-                $id = $_GET['id'];
-                try {
-                    $gestor->delete($id);
-                    echo "<script type='text/javascript'>alert('Se ha eliminado exitosamente al estudiante.');</script>";
+                    $gestor->insert($estudiante);
+                    echo "<script type='text/javascript'>alert('Registro de estudiante exitoso.');</script>";
                 } catch (Exception $e) {
                     echo "<script type='text/javascript'>console.log(`Error: $e`);</script>";
-                    echo "<script type='text/javascript'>alert('No se pudo eliminar al estudiante, lo sentimos.');</script>";
+                    echo "<script type='text/javascript'>alert('No se pudo realizar el registro, lo sentimos.');</script>";
                 }
                 echo "<script type='text/javascript'>window.location.href='../view/home.php';</script>";
                 break;
+            case 'update':
+                $id = $_POST['id'];
+                $estudiante = new Estudiante($id, $nombre1, $nombre2, $apellido1, $apellido2, $direccion, $telefono, $correo, true);
+                try {
+                    $gestor->update($estudiante);
+                    echo "<script type='text/javascript'>alert('Se actualizó con éxito al estudiante.');</script>";
+                } catch (Exception $e) {
+                    echo "<script type='text/javascript'>console.log(`Error: $e`);</script>";
+                    echo "<script type='text/javascript'>alert('No se pudo realizar la acción.');</script>";
+                }
+                echo "<script type='text/javascript'>window.location.href='../view/home.php';</script>";
+                break;
+        }
+
+        break;
+    case 'GET':
+        $action = $_GET['action'];
+        if ($action == 'deleteStudent') {
+            $id = $_GET['id'];
+            try {
+                $gestor->delete($id);
+                echo "<script type='text/javascript'>alert('Se ha eliminado exitosamente al estudiante.');</script>";
+            } catch (Exception $e) {
+                echo "<script type='text/javascript'>console.log(`Error: $e`);</script>";
+                echo "<script type='text/javascript'>alert('No se pudo eliminar al estudiante, lo sentimos.');</script>";
+            }
+            echo "<script type='text/javascript'>window.location.href='../view/home.php';</script>";
         }
 }
